+++ changelog.md
@@ -8,6 +8,7 @@
 * [added] Visible name for locations
 * [added] Extended edit options for locations
 * [added] Taxon statistics on list of mediafiles
+* [changed] Layout of the list of mediafiles
 
 ## 02-2024